Add Sort Method to CustomList using Quick Sort Algorithm
Quick Sort works by taking a pivot point, in our case always the last in a series, and testing if each value is higher and lower, moving it left or to the right of it in the array. They same thing is then done for those that are lower than the pivot and those that are higher than the pivot (as we need not test the pivot again because we know its in the right spot.)
This is recursive until pivot equals the series start and end.